requirements.txt
opencv-python
numpy
urx
pyyaml